refactor: PageTitle 컴포넌트를 제거하고 header에 인라인 처리
