+++ packages/cds-ui/src/components/index.ts
@@ -3,7 +3,6 @@ export { default as File } from './file/file';
 export { default as FloatingButton } from './floating-button/floating-button';
 export { default as ImageContainer } from './image-container/image-container';
 export { default as Modal } from './modal/modal';
-export { default as PageTitle } from './page-title/page-title';
 export { default as Tag } from './tag/tag';
 export { default as TextContent } from './text-content/text-content';
 export { default as Title } from './title/title';
