--- packages/cds-ui/src/components/page-title/page-title.tsx
@@ -1,17 +1,0 @@
-import * as styles from './page-title.css';
-
-interface PageTitleProps {
-  title: string;
-  count: number;
-}
-
-const PageTitle = ({ title, count }: PageTitleProps) => {
-  return (
-    <div className={styles.container}>
-      <h2 className={styles.title}>{title}</h2>
-      <span className={styles.count}>{count}</span>
-    </div>
-  );
-};
-
-export default PageTitle;
